restructure package bosehubbard into smaller modules
